pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pantheon-framework"
version = "0.1.0"
description = "Operating System for Containerized AI Teams - RAE Framework for structured AI workflows"
authors = [{name = "Pantheon Team", email = "team@pantheon-framework.dev"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
keywords = ["ai", "framework", "workflow", "automation", "rae", "retrieval-augmented-execution"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "PyYAML>=6.0",
    "jsonschema>=4.19",
    "Jinja2>=3.1",
    "jsonnet>=0.21",
    "click>=8.1",
    "pydantic>=2.4",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4",
    "pytest-mock>=3.12",
    "pytest-cov>=4.1",
]
dev = [
    "mypy>=1.5",
    "ruff==0.13.2",
    "pre-commit>=3.5",
    "types-PyYAML>=6.0",
    "types-jsonschema>=4.25",
    "types-jsonnet>=0.21",
]

[project.scripts]
pantheon = "pantheon.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["pantheon*"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
no_implicit_optional = true
no_implicit_reexport = true
show_error_codes = true
show_error_context = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "PIE", "SIM", "RET"]
ignore = ["E501"]  # Line length handled by formatter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["pantheon"]
force-sort-within-sections = true
split-on-trailing-comma = true